+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/opmodes/teleop/DashboardTeleOp.kt
@@ -1,4 +1,4 @@
-package org.firstinspires.ftc.teamcode.opmodes
+package org.firstinspires.ftc.teamcode.opmodes.teleop
 
 import com.acmerobotics.dashboard.FtcDashboard
 import com.qualcomm.robotcore.eventloop.opmode.Disabled
