+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/opmodes/teleop/SetAllianceRed.java
@@ -1,21 +1,20 @@
-package org.firstinspires.ftc.teamcode.opmodes;
+package org.firstinspires.ftc.teamcode.opmodes.teleop;
 
 import com.qualcomm.robotcore.eventloop.opmode.LinearOpMode;
 import com.qualcomm.robotcore.eventloop.opmode.TeleOp;
 
-import org.firstinspires.ftc.teamcode.subassemblies.Underglow;
 import org.firstinspires.ftc.teamcode.util.Global;
 
-@TeleOp(name = "Set Alliance Blue", group = "!set alliance")
-public class SetAllianceBlue extends LinearOpMode {
+@TeleOp(name = "Set Alliance Red", group = "!set alliance")
+public class SetAllianceRed extends LinearOpMode {
     @Override
     public void runOpMode() {
-        telemetry.addLine("Press START to set alliance to BLUE, or press STOP now to disable");
+        telemetry.addLine("Press START to set alliance to RED, or press STOP now to disable");
         telemetry.update();
         Global.alliance = null;
         waitForStart();
         if (opModeIsActive()) {
-            Global.alliance = Global.Alliance.BLUE;
+            Global.alliance = Global.Alliance.RED;
         }
         requestOpModeStop();
     }
