+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/opmodes/teleop/SimpleTeleOp.kt
@@ -1,4 +1,4 @@
-package org.firstinspires.ftc.teamcode.opmodes
+package org.firstinspires.ftc.teamcode.opmodes.teleop
 
 import com.acmerobotics.dashboard.FtcDashboard
 import com.acmerobotics.dashboard.telemetry.MultipleTelemetry
