+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/opmodes/teleop/UtilityCameraFrameCapture.java
@@ -31,7 +31,7 @@
  * THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
  */
 
-package org.firstinspires.ftc.teamcode.opmodes;
+package org.firstinspires.ftc.teamcode.opmodes.teleop;
 
 import android.util.Size;
